refactor: separa cálculos da correção técnica
